% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGNAL.R
\name{compute_cov}
\alias{compute_cov}
\title{Compute covariance matrix}
\usage{
compute_cov(X, G, block.size)
}
\arguments{
\item{X}{Input FBM object.}

\item{G}{Group ont-hot matrix.}

\item{block.size}{Block size.}
}
\description{
Compute covariance matrix
}
