% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGNAL.R
\name{group_mmt}
\alias{group_mmt}
\title{Compute the sum of the outer product of the scaled group means}
\usage{
group_mmt(X, G, output.all = FALSE)
}
\arguments{
\item{X}{Input FBM object.}

\item{G}{Group one-hot matrix.}

\item{output.all}{Whether to output all results.}
}
\description{
Compute the sum of the outer product of the scaled group means
}
