\name{Ma}
\alias{Ma}
\title{ Create a moving average (MA) model }
\description{
  Returns a moving average MA model. The model could represent a filter or system model.
}
\usage{
Ma(b)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{b}{ moving average (MA) polynomial coefficients }
}
\details{
}
\value{
  A list with the MA polynomial coefficients of class \code{Ma}.
}
\author{ Tom Short, EPRI Solutions, Inc., (\email{tshort@eprisolutions.com}) }
\seealso{ See also \code{\link{Arma}}
}

\keyword{ math }