% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGNAL.R
\name{Run.gcPCA}
\alias{Run.gcPCA}
\title{Group-centralized principal component analysis}
\usage{
Run.gcPCA(
  X,
  meta,
  g_factor,
  b_factor,
  lambda = 50,
  npcs = 30,
  excluded.cells = NULL,
  output.all = FALSE,
  do.scale = F,
  do.cosine = F,
  block.sizeGB = 2,
  block.size = 4000
)
}
\arguments{
\item{X}{Input data matrix.}

\item{meta}{Input metadata (data.frame).}

\item{g_factor}{Group variable (s).}

\item{b_factor}{Batch variable (s).}

\item{lambda}{Tuning parameter.}

\item{npcs}{How many dimensions to reduce.}

\item{excluded.cells}{Metadata about which cells should not be used.}

\item{output.all}{Whether to return all calculated values.}

\item{do.scale}{Whether to perform global scaling.}

\item{do.cosine}{Whether to cosine normalize the integration result.}

\item{block.sizeGB}{Memory settings for block computing.}

\item{block.size}{Block size.}
}
\description{
Group-centralized principal component analysis
}
