% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGNAL.R
\name{Run.LabelTransfer.Single}
\alias{Run.LabelTransfer.Single}
\title{Predict cell labels of query data using cell labels of reference data.}
\usage{
Run.LabelTransfer.Single(
  ref_data,
  ref_group,
  query_data,
  do.cosine = T,
  do.scale = F,
  conf.score = T,
  proj.dims = 30,
  eta = 0.25,
  smooth.k = 5
)
}
\arguments{
\item{ref_data}{Reference data matrix.}

\item{ref_group}{Reference labels.}

\item{query_data}{Query data matrix.}

\item{do.cosine}{Whether to cosine normalize the data matrices.}

\item{do.scale}{Whether to perform scaling.}

\item{conf.score}{Whether to calculate confidence scores.}

\item{proj.dims}{Projection dimension of query data to reference data.}

\item{eta}{The proportion of cells that are considered reliably projected.}

\item{smooth.k}{Using kNNs to smooth the scores.}
}
\description{
Predict cell labels of query data using cell labels of reference data.
}
