% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGNAL.R
\name{Run.PGTV.Single.Cycle}
\alias{Run.PGTV.Single.Cycle}
\title{Predict cell labels for a data matrix using its given cell labels.}
\usage{
Run.PGTV.Single.Cycle(
  data,
  group,
  do.cosine = T,
  do.scale = F,
  max.iters = 20,
  thresh = 1e-05
)
}
\arguments{
\item{data}{Data matrix.}

\item{group}{Given labels.}

\item{do.cosine}{Whether to cosine normalize the data matrices.}

\item{do.scale}{Stopping threshold for classification agreement metrics.}

\item{max.iters}{The maximum number of iterations.}

\item{thresh}{Whether to perform scaling.}
}
\description{
Predict cell labels for a data matrix using its given cell labels.
}
