% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIGNAL.R
\name{Run.PGTV.Single}
\alias{Run.PGTV.Single}
\title{Predict cell labels using group technial effects.}
\usage{
Run.PGTV.Single(ref_data, ref_group, query_data, do.cosine = T, do.scale = F)
}
\arguments{
\item{ref_data}{Reference data matrix.}

\item{ref_group}{Reference labels.}

\item{query_data}{Query data matrix.}

\item{do.cosine}{Whether to cosine normalize the data matrices.}

\item{do.scale}{Whether to perform scaling.}
}
\description{
Predict cell labels using group technial effects.
}
