% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trailing.R
\name{trailing}
\alias{trailing}
\title{Trailing}
\usage{
trailing(v, days = 28)
}
\arguments{
\item{days}{Number of days trend should cover. Defaults to 28.}
}
\description{
This function calculates the trailing trend in a vector.
}
\examples{
trailing()
}
\keyword{trailing}

